feat: calculatePortion 기능 추가(포트폴리오 상품별 비율 계산 기능)

+++ src/main/java/com/be/portfolio/dto/res/PortfolioPortionDto.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PortfolioPortionDto {
-    private int totalSaving;
-    private int totalBond;
-    private int totalFund;
-    private int totalStock;
+    private double totalSaving;
+    private double totalBond;
+    private double totalFund;
+    private double totalStock;
 }
